NHWC pooling: offset computation fix for large sizes

+++ src/cpu/nhwc_pooling.cpp
@@ -43,15 +43,15 @@ namespace cpu {
                                  : MEM_D(name).blocking_desc().strides[0][4];
 
 namespace nhwc_pooling {
-    size_t strided_offset(const int  _n, const int _sn,
-                          const int  _d, const int _sd,
-                          const int  _h, const int _sh,
-                          const int  _w, const int _sw)
+    size_t strided_offset(const int _n, const size_t _sn,
+                          const int _d, const size_t _sd,
+                          const int _h, const size_t _sh,
+                          const int _w, const size_t _sw)
     {
-        return   (size_t)_n * (size_t)_sn
-               + (size_t)_d * (size_t)_sd
-               + (size_t)_h * (size_t)_sh
-               + (size_t)_w * (size_t)_sw;
+        return   _n * _sn
+               + _d * _sd
+               + _h * _sh
+               + _w * _sw;
     }
 }
